Fix plugin file the structure and format

+++ demo/os/linux/easylogger/port/elog_port.c
@@ -32,7 +32,7 @@
 #include <unistd.h>
 #include <time.h>
 
-#include <elog_file.h>
+#include <file/elog_file.h>
 static pthread_mutex_t output_lock;
 
 /**
@@ -60,7 +60,7 @@ ElogErrCode elog_port_init(void) {
  */
 void elog_port_output(const char *log, size_t size) {
     /* output to terminal */
-    printf("%.*s", size, log);
+    printf("%.*s", (int)size, log);
 #ifdef ELOG_FILE_ENABLE
     /* write the file */
     elog_file_write(log, size);
